fix(e2e): extract fetchJson to a pure module
`wp-env-fixtures.js` calls `readCredentials()` at module scope, so
importing it anywhere without a provisioned wp-env throws. That made the
unit test for `fetchJson` unrunnable on CI, where
`.wp-env.credentials.json` does not exist — it only passed locally
because a stale credentials file happened to sit in the worktree.

Moving the helper to `e2e/fetch-json.js` keeps the retry logic testable
without dragging in the credential-bound fixture setup.

---
