+++ e2e/fetch-json.test.js
@@ -6,7 +6,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 /**
  * Internal dependencies
  */
-import { fetchJson } from './wp-env-fixtures';
+import { fetchJson } from './fetch-json';
 
 const CREDS = { authHeader: 'Basic dGVzdDp0ZXN0' };
 const URL = 'http://localhost/wp-json/wp-block-editor/v1/settings';
